+++ YesterdayNews/Views/Shared/_ArticlePartialView.cshtml
@@ -34,11 +34,6 @@
                 <h3 class="text-left mb-0">@Model.Category?.Name.ToUpper()</h3>
             </div>
         </div>
-        <div class="col-md-2">
-            <a asp-controller="Article" asp-action="Index" class="btn btn-outline-secondary form-control btn-sm">
-                <i class="bi bi-arrow-left me-2"></i>Back
-            </a>
-        </div>
     </div>
     </div>
 
@@ -57,7 +52,7 @@
             alt="@Model.Headline">
         }
     </div>
-    @if (!User.IsInRole(StaticConsts.Role_Customer))
+    @if (!User.IsInRole(StaticConsts.Role_Customer) && User.Identity.IsAuthenticated)
     {
 
         <span class="badge bg-@GetStatusColor(@Model.ArticleStatus.ToString()) px-2 px-md-3 py-1 py-md-2 fs-6 text-uppercase position-absolute"
@@ -69,7 +64,6 @@
 
 <!-- Headline -->
 <div class="text-center mb-3">
-
     <h1 class="display-5 fw-bold">@Model.Headline</h1>
 </div>
 
@@ -108,7 +102,7 @@
             //Using server side approach
             <div class="like-container">
 
-                <form method="post" asp-action="ToggleLike" asp-controller="Article" asp-route-id="@Model.Id" class="mb-0">
+                <form method="post" asp-action="ToggleLike" asp-controller="Home" asp-route-id="@Model.Id" class="mb-0">
                     <button type="submit" class="btn btn-sm btn-outline-light border-0 position-relative">
                         <i class="bi @(Model.IsLikedByCurrentUser ? "bi-heart-fill text-danger" : "bi-heart")"></i>
                         <span id="likesCount">@Model.Likes</span>
@@ -139,97 +133,3 @@
     </div>
 </div>
 
-<!-- Management Actions -->
-@if (!User.IsInRole(StaticConsts.Role_Customer) && Model.ArticleStatus.ToString() != StaticConsts.ArticleArchived)
-{
-    
-<div class="card border-0 shadow-sm my-5">
-        <div class="card-header bg-primary">
-            <h5 class="mb-0">Manage Article</h5>
-        </div>
-
-        <div class="card-body d-flex flex-wrap gap-2 justify-content-center">
-            @if ((User.IsInRole(StaticConsts.Role_Journalist) || User.IsInRole(StaticConsts.Role_Admin))  && Model.ArticleStatus.ToString() == StaticConsts.ArticleDraft)
-            {
-                <form asp-action="SubmitForReview" asp-route-id="@Model.Id" method="post">
-                    <button type="submit" class="btn btn-warning">
-                        <i class="bi bi-send-check"></i> Submit for Review
-                    </button>
-                </form>
-            }
-            @if ((User.IsInRole(StaticConsts.Role_Editor) ||
-            User.IsInRole(StaticConsts.Role_Admin)) && 
-            (Model.ArticleStatus.ToString() != StaticConsts.ArticleDraft 
-            && Model.ArticleStatus.ToString() == StaticConsts.ArticlePendingReview))
-            {
-                <form asp-action="Publish" asp-route-id="@Model.Id" method="post">
-                    <button type="submit" class="btn btn-success">
-                        <i class="bi bi-check-circle"></i> Publish
-                    </button>
-                </form>
-            }
-           
-            @if ((User.IsInRole(StaticConsts.Role_Editor) || 
-            User.IsInRole(StaticConsts.Role_Admin)) && Model.ArticleStatus.ToString() == StaticConsts.ArticlePendingReview)
-            {
-                <button type="button" class="btn btn-danger" data-bs-toggle="modal" data-bs-target="#rejectModal">
-                    <i class="bi bi-x-square"></i> Reject
-                </button>
-
-                <!-- Rejection Modal -->
-                <div class="modal fade" id="rejectModal" tabindex="-1" aria-labelledby="rejectModalLabel" aria-hidden="true">
-                    <div class="modal-dialog">
-                        <div class="modal-content">
-                            <div class="modal-header">
-                                <h5 class="modal-title" id="rejectModalLabel">Reject Article</h5>
-                                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                            </div>
-                            <form asp-action="Reject" asp-route-id="@Model.Id" method="post">
-                                <div class="modal-body">
-                                    <div class="mb-3">
-                                        <label for="rejectionReason" class="form-label">Reason for rejection:</label>
-                                        <textarea class="form-control" id="rejectionReason" name="rejectionReason" rows="3" required></textarea>
-                                    </div>
-                                </div>
-                                <div class="modal-footer">
-                                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-                                    <button type="submit" class="btn btn-danger">Confirm Rejection</button>
-                                </div>
-                            </form>
-                        </div>
-                    </div>
-                </div>
-
-            }
-          
-         
-            @if (User.IsInRole(StaticConsts.Role_Editor) || User.IsInRole(StaticConsts.Role_Admin))
-            {
-                <form asp-action="Archive" asp-route-id="@Model.Id" method="post">
-                    <button type="submit" class="btn btn-secondary">
-                        <i class="bi bi-archive"></i> Archive
-                    </button>
-                </form>
-            }
-            @if (!User.IsInRole(StaticConsts.Role_Customer))
-            {
-
-                <a role="button" href="/Article/Edit/@Model.Id" class="btn btn-outline-secondary border-primary">
-                    <i class="bi bi-pencil-square"></i> Edit
-
-                </a>
-
-
-            }
-        </div>
-
-        @if (Model.ArticleStatus.ToString() == StaticConsts.ArticleRejected && !string.IsNullOrEmpty(Model.RejectionReason))
-        {
-            <div class="alert alert-danger mt-3 text-center">
-                <h5>Rejection Reason:</h5>
-                <p>@Model.RejectionReason</p>
-            </div>
-        }
-
-    </div>
-}
